net_state: Remove `prop_list`
The original reason of introducing `prop_list` is to indicate which
value is defined by user instead of default value or merged value.
This is hack which require a lot of documentation and human effort on
code review.

We do not need to do this hack any more as we have MergedNetworkState
which contains 3 states now:
 * Desired state
 * Current state
 * Merged  state

For `NetworkState`, only `ovsdb`, `dns`, `hostname` do not support
partial editing, hence all of them are protected by `Option<T>` to
indicate whether user desired or not.

Signed-off-by: Gris Ge <fge@redhat.com>