nm: use interface name instead of UUID
Current nmstate code is using UUID when referring the controller/parent in
NM profile, this cause problem when the controller/parent is in-memory
profile which change UUID after reboot. The OpenShift is using in-memory
profile for `br-ex` and will recreate it after reboot. This prevent us
from modifying the ports of `br-ex`.

Using interface name for controller/parent reference would fix this problem.

Existing test cases are enough to test this change.

Signed-off-by: Gris Ge <fge@redhat.com>