route: automatically set lock-cwnd if cwnd is requested
According to `man ip-route`, cwnd is ignored if the lock flag is
not used. If the user sets cwnd, we can asume that they want it to be
useful somehow, so let's set the lock flag for it.

As far as I know, the lock-cwnd flag doesn't have any noticeable side
effect other than making effective the use of the cwnd value, but anyway
we emit an info message to let the user know that it is being implicitly
set.

This flag is part of the RouteMetric::Lock netlink attribute. For the
moment, nmstate doesn't allow to use any of these flags directly, we only
use it internally for this lock-cwnd flag. Because of this, keep it private
for the moment so we don't have to decide the API design for it yet.

Signed-off-by: Íñigo Huguet <ihuguet@redhat.com>