deprecation: Deprecate PrettyState class
The PrettyState class is deprecated as it is only used for test code. The output of libnmstate.show() is already sorted, making PrettyState unnecessary. This commit marks the PrettyState class as deprecated and provides a warning message to inform users about its deprecation.

Additionally, test cases related to PrettyState are removed as they are no longer relevant.

Signed-off-by: jona42-ui <jonathanthembo123@gmail.com>