commands: Alias for `apply`, `show` nmstate subcommands. (#2579)
This commit introduces aliases for the `apply` and `show` subcommands
in the `nmstate` cli tools. These aliases allow users to use shorter and more
intuitive commands for common actions, enhancing the usability and user
experience of the CLI interface. With this enhancement, users can now
use the `nmstatectl a` and `nmstatectl s` commands interchangeably with the
original `nmstatectl apply` and `nmstatectl show` commands, respectively.
This change simplifies command usage and reduces the cognitive load
on users, especially those who frequently interact with the `nmstate` cli
tools.

Signed-off-by: elomscansio <163124154+elomscansio@users.noreply.github.com>