test: integration: routes: fix wrong default table_id comparison
Nmstate uses the value '0' to indicate the default table ID, but the
kernel uses 254. Because of that, in assert_routes we must modify one of
the routes to compare them.

However, what we were doing was: if the desired route has table_id ==
default, then set the kernel route's table_id to default. This is wrong
because a desired route with table_id==default mustn't match with a
kernel route with table_id=123, for example.

Fix it by normalizing all routes to be compared to use the kernel's
default if and only if they were set to use the default table_id.

Signed-off-by: Íñigo Huguet <ihuguet@redhat.com>