test: integration: routes: fix tests checking for missing routes
Some tests were using assert_routes with an empty list of routes to
check that the routes has been removed. This is not correct because what
assert_routes does is to check that all the routes from the list are
present in the current network state. If the list is empty, it always
passes.

Use the new assert_routes_missing instead.

Signed-off-by: Íñigo Huguet <ihuguet@redhat.com>