setup.py: Add automatic installation of nmstate C library
This commit adds functionality to the setup.py script to automatically install the nmstate C library if it's not found on the system during installation. This enhancement ensures a smoother installation process for users by handling the dependency on the nmstate C library automatically.

The script checks for the presence of the nmstate package, and if it's not found, it attempts to install the C library using the install_nmstate_lib() function. If the installation fails, an error message is printed to the console, guiding the user to install the nmstate C library separately before retrying the installation of the Python package.

Signed-off-by: jona42-ui <jonathanthembo123@gmail.com>