route: Normalize route table of current route when determining route removed
Normalizing the `None` route table ID into default table ID when
creating the merged routes will cause the routes difference in new NM
conncetion and activated NM connection when applying the same desired
state the second time, as the result, the activated connection will be
dactivated first and activated again. This is troublesome, because the
routes added by iproute2 utility will be deleted by mistake. To fix
that, normalize the route table of the current routes in the activated
NM connection before comparing the difference between current routes
and new routes.

Resolves: https://issues.redhat.com/browse/RHEL-29241

Signed-off-by: Wen Liang <liangwen12year@gmail.com>