markdown: Correct badges and fix markdown errors
Fixed the CI badge to link to the CI status page. Corrected the crates.io badge to link to the nmstate crate page. Also, fixed the docs.rs badge to link to the nmstate documentation page.

With these corrections, the badges now correctly link to the appropriate pages, and markdown errors have been rectified.

Signed-off-by: jona42-ui <jonathanthembo123@gmail.com>