tests: add centos mirror list to avoid checksum errors
Let's add mirror lists for CentOS AppStream/BaseOS repos to prevent
Errors during downloading metadata for repository 'baseos' issues.
This is very likely caused by a repo resync and should not happen in
all other repos. So with this approach we are going to find working
repo rather than failing in the first checksum failure.