nm: Fix incorrect pre-deactivation on route changes
With route next-hop-address pointing to external(no in the CIDR of
interface IP), NM will automatically create direct route to this
next-hop-address. When applying this YAML with route desired but not
changed, nmstate incorrect treat it as route removed which lead to
full deactivation on the interface.

The root cause is `is_nm_ip_route_removed()` treat `Some(0.0.0.0)` !=
`None` for next-hop-address.

Both linux kernel and iproute tool are discarding this all-zero address
both for querying and changing.

To fix this issue:
 * Change `RouteEntry.sanitize()` to change all-zero address of
   next-hop-address to `None`.
 * Convert `NmIpRoute` to `RouteEntry`, so we can re-use `RouteEntry.is_match()`
   to tell whether route is changed or not in `is_nm_ip_route_removed()`.

Integration test case included.

Signed-off-by: Gris Ge <fge@redhat.com>