rust: Migrate zbus v1 to V5
Fedora 42 is removing support of zbus V1 and we are suggested to use V4
or V5.

The zbus V4 does not have Clone trait for OwnedValue which require
massive changes to our code base. Hence using zbus V5.

The zbus V5 need at least 1.77 to compile as it use a nightly feature of
rust https://github.com/rust-lang/rust/issues/91345 which only become
stable in rust 1.77.

It is impossible to compile this project in Rust 1.66 now, removing
CI test for compiling on Rust 1.66. Downstream release should
considering using Rust 1.77+ to compile this project.

Signed-off-by: Gris Ge <fge@redhat.com>