ovsdb: return an error if ovsdb socket is not available when applying
Return an error instead of warning when the desired state contains OVS
interfaces but Nmstate is not able to connect to the OVS daemon socket
and verification is not disabled.

If NetworkManager is in a different namespace with the OVS socket
mounted, then Nmstate could apply the configuration correctly but it
will fail the verification. Therefore, this operation should only be
allowed with `--no-verify`.

```
[2025-02-20T10:47:57Z INFO  nmstatectl] Nmstate version: 2.2.41
[2025-02-20T10:47:58Z ERROR nmstate::query_apply::net_state] PluginFailure: Desired state contains OVS interfaces, but not able to connect OVS daemon at socket /run/openvswitch/db.sock
NmstateError: PluginFailure: Desired state contains OVS interfaces, but not able to connect OVS daemon at socket /run/openvswitch/db.sock
```

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>