nm vlan: Support nm connection with empty `connection.interface-name`
When NM VLAN connection with empty `connection.interface-name`, NM will
use `parent.vlan_id` as name of this VLAN. With this setup, adding
routes to this vlan will trigger error in nmstate:

    NmstateError: InvalidArgument: Connection(MissingSetting):vlan:
    setting required for connection of type 'vlan'

This is because nmstate fail to treat this special activated NM connection
as exiting NM connection for desired VLAN interface, when creating it,
nmstate did not passing VLAN setup because nmstate thinks it already
exists.

To fix this issue, patch `get_exist_profile()` to check real
interface name of NmActiveConnection when not found such info from
NmConnection.

When search existing NM connection, we ignore connection with
`multi-connect: multiple|manual-multiple`.

Integration test case include and marked as tier1.

Resolves: https://issues.redhat.com/browse/RHEL-82100

Signed-off-by: Gris Ge <fge@redhat.com>