route rule: ensure idempotency for USE_DEFAULT_PRIORITY
When the desired state has route rules without a priority defined,
nmstate creates them with increasing values of priority. However,
this was causing that applying the same desired state twice created
the rule twice with different priorities. As a result, the operation
was not idempotent.

Fix it by searching for a matching rule in the current state and, if
there is any, use its priority value to fill the priority value of the
desired rule if it's empty. As a result, new rules are not created if
one that matches already exists.

Fix also a helper function of the integration tests to make it asserting
that the rule is not duplicated.

Signed-off-by: Íñigo Huguet <ihuguet@redhat.com>