nm: route rules: don't duplicate rules in NM profile
Each time that a routing rule was applied it was appended to the routing
rules of the NM profile. If it already existed in the profile, it was
added again, resulting in duplicated rules. As the kernel does not allow
duplicated rules, they were effectively applied only once, though.
Nonetheless, we shouldn't duplicate again and again the same rule in the
NM profile.

Before generating the NMConnection, internally we have to assign the routing
rules from the merged state to the different interfaces, as it is needed
by NetworkManager. First, the rules already existing in NM connections
are retrieved and stored in the field `BaseInterface::ipv4/6::rules`.
Then, we append there the new rules from the merged state (this is the
point where the duplication was happening). Finally, we generate the
NMConnection's routing rules from those `InterfaceIPv4/6::rules`.

Fix that by converting that `InterfaceIPv4/6::rules` from Vec to
HashSet, so no duplication can happen.

As I had to adapt the unit tests, I refactored them a bit to make
them more readable.

Signed-off-by: Íñigo Huguet <ihuguet@redhat.com>