nm: Fix HSR support
Support handling HSR interface type when generating `BaseInterface` from
`NmDevice`.

This was not a problem because old patch ws using `NmConnection` to
build up `BaseInterface`. When we refactor the NmConnection search code,
we unified the effort by using NmDevice only for querying.

Existing HSR test cases are enough for testing this patch.

Signed-off-by: Gris Ge <fge@redhat.com>