nm: implement From trait between InterfaceType and NmIfaceType
Previously, the `nm_dbus` was designed to be standalone crate without
dependency on nmstate, so `iface_type_to_nm()` is used to convert from
`InterfaceType` to `NmIfaceType`. But we don't have plan to release
`nm_dbus` crate anymore, so we should use From trait instead of function
for conversion between `InterfaceType` and `NmIfaceType`.

Signed-off-by: Gris Ge <fge@redhat.com>