route: detect nexthop conflicts and raise error
Currently, when a user tries to create a route which
already exists, but has a different nexthop, nmstate
returns a cryptic error:

Got netlink unknown error: code -22, msg: Invalid argument (os error 22)

When this happens, it is unclear whether the user
wants to create a new route, or alter existing
routes. Raise an error explaining what needs to
be changed by the user.

Resolves: https://issues.redhat.com/browse/NMT-1537
Signed-off-by: Jan Vaclav <jvaclav@redhat.com>