nm: Bump minimum supported NetworkManager version 1.46
Bumping minimum supported NetworkManager version to 1.46 which is
shipped by RHEL 9.4 and Ubuntu 24.04.

Removed the workaround in 802.1Q VLAN for NM 1.41-.

Did not remove other workaround codes because they will cause nmstate
not working in NM 1.46- at all.

Since the RHEL 8.10 is shipping NM 1.40, also removed the code for
running the tests in RHEL 8.

Assuming NM version for test is 1.46+, hence removed all the
conditional skip checker code for 1.46.

Changed github CI to run test on 1.46.

Signed-off-by: Gris Ge <fge@redhat.com>