dns: don't add DNS search and/or options to down NM connection
When configuring only search and/or options in dns-resolver we add them
to a NM connection and not to NM's global config, but nmstate may choose
a down interface, which results on errors at apply time or even on the
configurations being accepted but stay unused by NetworkManager.

Fix that by always checking that the interface is valid for DNS, and by
adding the requirement of the interface being up.

Signed-off-by: Íñigo Huguet <ihuguet@riseup.net>