Fix interface state detection in kernel mode
Any device with administrative state UP was being considered as `state:
up` in kernel mode. This is not correct because if the operational state
is not Up, for example because of the cable being disconnected, nobody
would think that the device is **really** up.

Fix it by ignoring the administrative state and consider only the
operational state.

Fixes https://github.com/nmstate/nmstate/issues/2798

Signed-off-by: Íñigo Huguet <ihuguet@riseup.net>