nm dns: interpret global-dns correctly after changes in NM
In the past, NetworkManager ignored all DNS configurations from
connections if there was any global DNS configuration. This was changed in
NM 1.44, that started to allow defining a global-dns section without
defining any global-dns-domain-* section.

After this change, betwen NM 1.44 and 1.56, searches and options from
global-dns were merged with the per-connection ones, but due to a bug,
not in a consistent way. NM 1.56 fixed this bug but also reverted the
behavior of merging global and per-connection searches and options, as
it didn't make much sense and was not intuitive.

The result:
- Before NM 1.44, global-dns and global-dns-domain-* had to be defined
  either both or none. They overwrited per-connection DNS configs.
- From NM 1.44 to 1.54: global-dns can be defined without a
  global-dns-domain-*, and it's merged with per-connection configs, but
  not always due to a bug.
- Since NM 1.56:
  - If only global-dns is specified, NM uses searches and options from
    global DNS and nameservers from connections.
  - If global-dns-domain-* is defined, NM uses everything from global
    DNS. If global-dns is not defined, it is assumed that it exists but
    empty.

This commit makes nmstate to read DNS configuration as it behaves in NM
>= 1.56. We only support NM >= 1.46 so we don't need to worry about the
behavior in NM <= 1.44. For versions 1.46-1.54, nmstate will write the
configurations correctly, because we always purge global and
per-interface DNS configs before applying the new one; but we may make
some small mistakes at reading a configuration not written by nmstate.
Anyway, this is unlikely to happen and it will be fixed after a write.
Also, the NM fix will be likely backported at least up to NM 1.46, so we
won't even make this reading error.

Signed-off-by: Íñigo Huguet <ihuguet@riseup.net>