nm dns: differentiate empty global DNS from undefined one
The distinction is important because an empty [global-dns] overwrites
the configs from connections, but if [global-dns] is not defined at all
then it doesn't.

Important: this commit depends on NM's commit "core: dns: show in D-Bus
if [global-dns] is defined but empty", so it needs a NM version with
that fix applied.

Signed-off-by: Íñigo Huguet <ihuguet@riseup.net>