policy route: Fix problem when removing routes via policy
Nmstate ignore policy to remove routes using not-interface matching, for
example:

```yml
desired:
  routes:
    config:
      - destination: 203.0.113.0/24
        state: absent
```

This is because `NetworkPolicy` use `MergedRoutes.get_diff()` to
generate `NetworkState`, and it has it own implementation of applying
absent routes which does not handle route match correctly.

To fix the issue, instead of re-calculate the changed routes, we use
`MergedRoutes.changed_routes` which is well tested by apply action.

Unit test and integration test case included.

Signed-off-by: Gris Ge <fge@redhat.com>