service: Support `.yaml` extension
Introduce the support of file named with `.yaml` extension for desired
state file in `/etc/nmstate` folder.

When two files holding the same name but one with `.yml` and another
with `.yaml`, nmstate.service will fail with error message indicating
the confliction. The reason of raising error for this use case is
because both files will be renamed to `foo.applied`. Changing the
renamed file to `foo.yaml.applied` will be change of behavior which is
too risky comparing the corner case of having two YAMLs holding the same
name with different extensions.

Integration test case included.

Signed-off-by: Gris Ge <fge@redhat.com>