ipsec: New option `nm-auto-defaults: false`
Introducing new ipsec option `nm-auto-defaults: true|false`, default to
true. When set to false, instruct NetworkManager to not set default
values but pass only desired values to ipsec libreswan daemon as it is.

With this, host-to-host(P2P) mode is not required to set `leftmodecfgclient: no`
explicitly which align with `ipsec.conf` manpage of libreswan.

It is recommend to set `nm-auto-defaults: false` and perform following
steps to align your YAML with libreswan daemon default values:
 * Upgrade NetworkManager-libreswan to 1.2.28 or later version.
 * Set `leftmodecfgclient: yes` explicitly if you need pull config from
   remote.
 * Add `@` to `leftid` and `rightid` if you don't want it been resolved
   to IP address. (e.g. `leftid: '@cli-a.example.org'`).
 * Define `rightsubnet` explicitly. (e.g.`rightsubnet: 0.0.0.0/0`)

Resolves: https://issues.redhat.com/browse/RHEL-26350

Signed-off-by: Gris Ge <fge@redhat.com>