nm: deactivate hsr interface before reactivating when changed
NM has special handling for "virtual" interfaces, and does not
remove them from kernel when connections are changed. This is
a problem because now-modified properties are not properly
applied to such interfaces.

HSR is one of such virtual interfaces, and since it does
not have a `changelink` handler in kernel, implementing
reapply is out of the question. Thus, we have no other
option but to deactivate the interface, to force NM into
recreating it from scratch.

Resolves: https://issues.redhat.com/browse/RHEL-100773

Signed-off-by: Jan Vaclav <jvaclav@redhat.com>