nispor: consider re-ordering when assigning static ip
Fixes Pull Request #2953

The original pull request fixed the problem of new ip address appending
instead of updating. This was done by removing ip address in current state
that are in the desired state and adding new ip addresses.

However, if the ip addresses are re-ordered in the new state, the
previous logic will not update the ordering. The new logic will delete
current ip address if it does not exist in desired state or the order
has changed.

Example case:

Old state
```
        name: lo
        type: loopback
        ipv4:
          enabled: true
          address:
          - ip: 127.0.0.1
            prefix-length: 8
          - ip: 127.0.0.2
            prefix-length: 32
        ipv6:
          enabled: true
          address:
          - ip: ::1
            prefix-length: 128
          - ip: ::2
            prefix-length: 128
```
New state
```
        name: lo
        type: loopback
        ipv4:
          enabled: true
          address:
          - ip: 127.0.0.2
            prefix-length: 32
          - ip: 127.0.0.1
            prefix-length: 8
        ipv6:
          enabled: true
          address:
          - ip: ::2
            prefix-length: 128
          - ip: ::1
            prefix-length: 128
```

Signed-off-by: Rahul Rajesh <rajeshrah22@gmail.com>